Better inject reset handling
This change pulls the nested-reset mixin into the annotator-frame class
and then increases the specificity of the rest of the styles underneath
the frame by giving them descendent selectors of the annotator-frame.

Additionally, to avoid interfering with site styles, our stylesheet of
resets (reset.scss) is no longer included in common.scss, but imported
directly under the annotator-frame class.

Fix #1366