Use `request.domain` as the local auth provider
Rather than using `request.server_name`, which may or may not be
reliable, depending on the configuration, use the domain property
of the request which will be just the hostname part of the HTTP host
header with fallback to the server name.

Protect your production applications behind a trusted reverse proxy!

This is required after a behavior change with respect to the server
name in Gunicorn 19 but it also fixes the provider string on dokku.
All around I think it's less surprising.