Update all account management forms for consistency
Now all forms use the profile view name plus "form". For example
the form for the edit_profile view would be editProfileForm. Each
ngModel is namespaced on the scope by the form name. So again this
would be $scope.editProfile. Lastly each field name uses the
corresponding field in the request payload. All fields are camel case.

This should make it far easier in future to read a form and match it
to the http request.