Remove heatmap tab blinking and dead styles
I find the tab blinking to be too subtle to be meaningful and not
a worthy enough feature to justify more jQuery UI use or storing
an additional property on the annotation (see #1207).

If we want a feature like this we'll have to come back and revisit
it with a more critical eye.