One last time, changing the default secret size
So much misinformation around the net. I was first led to believe
by the WebOb documentation (through the Pyramid documentation) that
a secret of 128 bits was a good idea as it matches the internal
block size of the hash function.

The docs also say "64 characters". I don't know of any definition of
character that results in a 64 character string being 128 bits. Nor
is the block size of the default hash (SHA512) 128 bits; it is 1024.

Sigh.