Restore reloadOnSearch: false for viewer
There seems to be some combination of conditions which cause Firefox
to infinite loop on the $location service's watchers. Bisecting pointed
to the commit where I removed this. For now, let's slap this bandaid
on. The proper fix probably has to do with making the parameters we
pass to the iframe in its src actually do something (like validating
the origin of frames that connect to the sidebar) and then clearing the
parameters from the URL so they don't cause a reload after the initial
redirection to /viewer.