Drop heatmap comment tab and convert tabs to css
Kicking off work on revamping the heatmap a little bit. This change
removes the comment tab. Clearly, something needs to be done to restore
discoverability to comments, but there seemed to be some consensus that
the comment tab was not comprehensible enough.

Additionally, I got pretty sick of the tabs bouncing around. I removed
the transitions, slimmed the tabs down a litle bit, and remade them in
pure CSS. The pure CSS has the advantage of avoiding extra network
requests to fetch the tab SVGs. Also, the SVGs contained some padding
and borders that is easier to account for if its all in the CSS.