Remove bogus reference in socket.coffee
The way our coffee is being bundled is actually problematic because
each of the individual files insn't wrapped in an anonymous function,
only the whole coffee bundle is. Therefore, the configure reference
in socket.coffee should have failed. It doesn't. Who knows what else
might be confused as a result of this situation. I'll sort it all out
this week.