Add back PDF.js to Chrome extension
This commit contains all the actual changes to H,
which are required for the new PDF support.

1. Use different insertion method.

Old method:
 - First PDF.js does it's magic, and redirects the page from the
   original document to it's own HTML content (the viewer)
 - Then embed.js injects the Hypothes.is code to the HTML page.

New method:
 - We embed our H code directly in the PDF.js viewer.
   When we want to activate H, we redirect the browser
   to the viewer.

For the integrated viewer, I have created a different embed.js version,
called embed-pdf.js, tailored for this specific use-case. The main
difference is that it actually uses eval() to run the code,
instead of inserting a script tag.

This was necessary because Chrome really, really hates inline
JS code in extensions, and it's probably impossible to configure
a CSP which would allow this.

2. Don't load the Document plugin for PDF files

3. Add support for proper URL and metadata extraction for PDF