Only show the citation on the standalone annotation
This fixes #1390. A new controller for the standalone annotation
allows us to determine how to render the annotation and this information
is passed down to the AnnotationController via a combination of scope
properties and attributes.

There are now two distinct states for an annotation, either embedded in
which the annotation is rendered on top of the original document or
standalone, where it appears separately from its document.