Prune empties in threading plugin
I had decided not to do this initially in order to show annotations
that were missing. We'll have to decide how we want to show the holes,
but it's much better if we prune the containers that aren't used so
that we don't end up with empty threads sitting around all over the
place.

The only tricky thing is that when new annotations are loaded we must
thread everything again or existing containers that were lifted up to
the root will not attach themselves to newly loaded parents.