Don't read the parent when sending stream updates
There's no reason to do this. Not only is it technically incorrect
to insert the parent's text as the quote of the child (if we want
that we should be doing it when we create the child) it's also not
storing it as a target but in a top level 'quote' property.

Fix #1264
Related to #1207