Move share and reply back inside the annotation
Move the edit and delete actions back down to the bottom of the
annotation alongside the share and reply actions.

Improve the experience when searching and composing by adding a
freeze function to the thread filter so that edits can freeze their
filter in a deactivated state in order to always match the filter.
This change keeps edits visible all the time and keeps the reply
and show more counts accurate.

Make the annotation and thread markup a bit more semantic and use
this to tweak some margins now that the reply count is below the
actions.

Drop the "<nobody>" placeholder for unattributed annotations and
instead just opt not to show the username or privacy and highlight
indicators.

Change the markdown directive to use an attribute and tighten up
the collapsed thread styles.