Tweak logging setup
Change the logging setup to use the same format as Gunicorn defaults.
Strip the redundant default args to the console handler. Remove the
extra loggers whose defaults are fine. Remove the access log config.
Set the root logger in development.ini to INFO.

I had a fundamental misunderstanding of Python loggers. I thought that
for a logger to handle a message its effective level was determined by
traversing all the way to the root and clamping the log level to the
highest configured logger. It turns out that it's the first ancestor
to have a level set that determines the effective level. Therefore,
it's sufficient to set 'h' to DEBUG but leave the root at INFO in
order to get debug message from the application and info messages
from other packages.

The gunicorn.access logger is unnecessary to configure. It isn't
enabled without a switch in the server config, or a --access-logfile,
so we'll leave it out.

I also removed places where I wrote WARN instead of WARNING (all of
these were redundant since NOTSET is the default and WARNING is the
default for the root) or had the propagate setting redundantly set to
True (also the default).

Close #1512