Refactor includes to make all the models
Rather than putting the notifications module into the multiauth
configuration, which was a nasty hack to get it to load early,
remove the automatic inclusion of pyramid_basemodel and put it into
the load order after h in `pyramid.includes`. That gives multiauth
time to include the local account and its models before the
pyramid_basemodel is included (at which point the models are actually
created).