Virtual root index support for API
This makes the API have an additional route that matches '/' as the
path, taking precedence over the normal index page whenever the
containment includes the APIResource. With this change, a request to
the root with the X-Vhm-Root header set to "/api" will return the API
service document, enabling a separate subdomain for API requests.