tests: Only run tests from the h package
By default, our bootstrap script installs the virtualenv on top of the
repository root. When we "pip install -e .", a copy of the source is
placed into the "local/" directory at the repository root[1], which
causes tests to be discovered there, as well.

We can't simply ignore the "local/" directory, because py.test's
norecursedirs option only allows patterns to be specified[2], not
absolute paths, and including "local" would cause tests in
"h.auth.local" to be skipped.

This commit also removes a few other ignores from the norecursedirs
option that are no longer necessary as a result of the change to
setup.py.

[1]: https://github.com/pypa/virtualenv/blob/93cfa83/virtualenv.py#L1520
[2]: http://pytest.org/latest/customize.html#confval-norecursedirs