Avoid unnecessary annotation renders
The pattern of comparing the new value passed to a watcher with the old
value is idiomatic angular for avoiding the first digest. Instead, the
first digest is covered by the watch on the model reference identity.
After the initial rendering, the model will not change but the other
properties may, at which point re-rendering will occur.