Sticky privacy on the browser
Add angular-cache. The privacy directive gets a new attribute,
'privacy-default', which can be set to "false" to override the
default, stored privacy preference. Any other value or the absence
of the tag indicates that the privacy should default to whatever
value is stored. The stored value is changed whenever the user
explicitly sets the privacy level.