Fix double loading on install
This issue arises because we're not checking the status of the tab
before trying to inject. So when during installation a new tab is
created and the extension activated this triggers the change from
'inactive' -> 'active' in TabState which embeds the sidebar into the
loading tab. Later the update function is called when the tab loading is
complete injecting the sidebar a second time.

We now explicitly check the tab status in the injection handler to
prevent injection into a loading tab.

Fixes #1756.