Promote the annotation on the standalone page
The annotation loaded in the AnnotationViewerController is now promoted
to the top level of the viewer. This ensures that the parents are not
rendered.

We now also explicitly specify the top level threads as root threads
by passing the thread-root attribute to the directive. This allows
us to style what we would normally consider reply annotations as
document annotations.

Lastly, we now watch the thread attribute for changes to pick up the
changes to the threading.root property. Up until now it was set once
in the directives render call.