jinja_extensions: better include_raw content caching
The include_raw tag provided by the previous iteration of this
extension adds a constant value to the Jinja parse tree, but is unable
to tell Jinja that a particular file has been read in order to generate
it. This means that the pyramid autoreloader doesn't know to reread the
files in question when they change, making development that touches the
Angular templates rather tedious -- a manual server reload is required
every time a template is changed.

This commit changes the include_raw tag to an include_raw function,
which rereads the Angular templates on each rendering of app.html in
development mode, but memoizes the contents when in production.