Fix replies toggle visual glitch in edit mode
This fixes an issue where the replies button overlaps the license
message when the editor is switched into edit mode. This is achieved
by only applying the negative margin on the thread-reply element when
the annotation is in "view" mode.

An on on-mode-change handler passed to the annotation directive which is
called when the view mode changes. A class applying the margin is then
toggled on the reply element.