Zero-downtime deploys on dokku
\o/

It is important to ensure that any side effects of these first requests
appropriately handled. For instance, these checks will be made against
localhost and so assets generated from these requests will have bad
URLs. To fix this, ensure that the assets are built using the task in
the Procfile and the dokku-app-predeploy-tasks plugin is used to
generate the assets with relative paths first.