Remove negative margin on thread reply counts
Instead of using a negative margin to pull the reply count up in line
with the annotation actions, drop the clearfix on the annotation itself
so that the actions don't take up vertical layout space. Restore the
assurance that the actions and reply count don't get too close to the
card edge by putting the clearfix on the thread instead. Also, ensure
consistent use of the annotation-section class to get regular padding
rhythm on the annotation sections.

Fix #1808