JWT Bearer token support
Change the authorization code to issue and expect standard JWT bearer
tokens in the X-Annotator-Auth-Token header (for compatibility) or as
the HTTP Authorization header with the Bearer type.

Start accepting JWT Bearer tokens as authorization grants. When used
in this manner, the issuing client must already be authorized for the
subject. Currently, the only valid issuer is the built-in web client,
which is always pre-authorized. In the future, it will be necessary to
store validations in a database and only allow the JWT Bearer Token
flow generate tokens for users that have already completed another
flow with the client that allows for the use of refresh tokens. In
essence, the JWT Bearer Token is a self-issued, pre-authorized refresh
token when used as an authorization grant.