config: Allow setting feature flags from env
This commit makes it possible to override the value of feature toggles
from the system environment. Currently it's not possible to change these
without redeploying due to the inclusion of production.ini in this
directory, which undermines some of the flexibility of feature flags.

For example, to override a feature named 'trampoline' which is off by
default, you can set the environment variable

    FEATURE_TRAMPOLINE=True

Likewise, to disable a feature which is on by default:

    FEATURE_DONT_CRASH=False