Add normalised user id column (uid)
Both vanilla horus and our own custom user model had integrity bugs
relating to username normalisation.

Horus allows the creation of users with usernames differing only in
case. This wouldn't be a problem but for the default implementation of
`User.get_by_username`, which does case normalisation. This means that
if both "Bob" and "bob" signed up, only one of them would be able to
sign in (depending on the nature of the underlying database index).

The problem was compounded in `h` by a normalisation for the presence of
dots ('.') in usernames. It was possible to create multiple users whose
usernames differed only by dots, but it was only possible to retrieve
one of them from the database.

This commit adds a new column to the `user` table, `uid`, which contains
the normalised version of the username (lowercase, without dots) and has
a unique constraint. It is set whenever the username is set, including
at construction time.