Validate username uniqueness half-correctly
Rather than inventing a different way of checking to see if a user
exists, the register form now uses the same method as everything else,
namely

    User.get_by_username

Unfortunately, this is still at least partially broken because of the
bizarre assumptions that horus/deform makes, which is that it's okay to
do your registration in two stages:

    1. validate input
    2. insert into database

This doesn't work in the real world, where people can make concurrent
registration requests (either multiple people, or just double-tapping a
submit button). In the situation where requests A and B try to register
the same username at the same time, it's possible that these operations
happen in this order:

    A: SELECT (validation)
    B: SELECT
    A: INSERT
    B: INSERT

In this scenario, the last operation will explode and throw a
`sqlalchemy.exc.IntegrityError`. I have tried to do this properly, but
the shenanigans involved in correctly returning errors via horus/deform
seem absurdly complicated. Perhaps I'm missing something. I probably am.

Anyway, this completes the changes required to prevent bad data (i.e.
duplicate usernames) in our database, even if it doesn't cover all
cases.