Simplify sentry code and catch framework errors
Simplify the Sentry code by handling the exception catching in a
tween. By placing the tween underneath the exception view it is
not necessary for the exception view to know about Sentry.

Framework errors that occur above the exception view are now also
caught by adding Sentry as middleware in the pipeline.

The logger can be configured in the .ini file as well, reducing the
code for sentry to next to nothing.

Rather than having a registry setting for Sentry, use the environment
variable, 'SENTRY_DSN', as the only way to configure Sentry.