Centralise remember functionality for login
Move all the functionality that enables session persistence on login
(and autologin, when enabled for events like password reset and
registration) into h.accounts.subscribers.

Although this is primarily a simplification, it also fixes a bug where
if a logged-in user visits an old reset password page, the application
throws an exception.