Relieve the thread directive of list responsibilities
The thread directive is doing rather a lot:

- rendering a list of cards
- rendering a card
- rendering a reply within a card

Partially as a result, it's a rather unusual directive which doesn't
have an associated template and uses ng-include instead to pull in the
relevant HTML for the relevant pieces.

This change relieves the thread directive of the first of these
responsibilities, by making the rendering of the top-level list of cards
the job of the threadList directive, which has isolate scope and thus
now bounds the complexity of the underlying thread directives: only
those pieces of scope explicitly bound into the threadList directive are
available lower down the tree.

Doing this makes it possible for both the threadList and the thread
directives to have their own templates, and paves the way for more
simplification of the thread directive and controller.