Implement a claim flow for registered accounts
This commit adds a pair of routes which allow users who have been issued
claim tokens (or claim URLs, more accurately) to set passwords for the
accounts that they have reserved.

Any claim token which seems invalid or cannot be resolved to a user in
our database will result in a 404.

Users must be logged-out to claim accounts, and the account referenced
in the token cannot have been claimed before (i.e. it must have a blank
password field).