Make "collapsed" refer only to a thread's replies
The meaning of "collapsed" in the context of a thread was immensely
confusing:

- on a top-level annotation card it referred to whether replies were or
  were not shown
- on a reply, it referred to whether or not the entire annotation was
  hidden

This change removes all reference to the latter meaning. Annotations
(and replies) are still hidden appropriately when searches are
performed, but now collapsing and uncollapsing refers only to the
visibility of the *replies* of the addressed thread.