Fix real-time delete
When receiving a real-time notification about the removal
of an annotation, we emit an annotationDeleted event,
so that all components can react.

Earlier, we emited this event simply with the Annotation
bject that arrived on the wire.

However, some components couldn't deal with this, because
they were expecting to see the 'real' Annotation object,
the ones they already knew about.

So now we do a lookup before emiting the event, and use
the locally found objects instead.