Move accounts behind a feature flag
An unfortunate side effect of embedding the API implicitly as a
separate Pyramid app when the 'api' feature is turned on is that
we have no idea which includes to pass to this sub-app. Sidestep
the issue for the moment by just removing the one default include
we know to be totally inappropriate for inclusion in the API app:
the accounts system.