Add browserify config to package.json
This includes files that need to be shimmed under the "browserify-shim"
key. Annotator plugins that have no dependencies do not need to be
shimmed at the moment as they just attach themselves to the global
Annotator object.

jQuery is assigned as a global, this means that it is assumed to be
included separately and bound to the global scope.

Both jQuery and Annotator have aliases under the "browsers" key to
make them easier to reference.