Use key derivation to provide secret keys
Rather than asking people deploying h to provide new secret keys for
every distinct use of a secret in the application, use an HMAC-based key
derivation to generate these keys.

This simplifies deployment, and provides stronger guarantees of security
in the face of lazy configurations (such as setting all secrets the
same).

Capability URLs generated previously by the `h.notification` package
will continue to work as of this change due to a fallback option when
deserializing tokens. However, it is important to note that setting up
such a fallback for the session secrets is infeasible, and so all
current sessions will be invalidated by the deployment of this change.

Lastly, this change ensures that in the event of a misconfiguration
where a secret key is not provided, the application will issue a warning
and generate a transient key (from the OS PRNG).