Configure pyramid_basemodel in h.notification
In the API application, h.notification is included at pyramid boot time,
while h.accounts is not. This means that when the
h.notifications.Subscriptions model tries to query the database, it
explodes because there is no active session.

This change makes it possible to load h.notification and h.account in
either order.