Enable h.notification in the API application
The h.notification package contains subscribers for
h.events.AnnotationEvent, which send reply notification emails when new
replies are created.

Because the API is now an isolated WSGI application, it does not share a
registry with the main application. As such, we need to explicitly
include the notification package in the API application so that its
subscribers are correctly called when annotations are created through
the API.

This restores email notification functionality broken by 29c1500.