Commit the current transaction when processing messages
pyramid_mailer's default mailer won't actually send messages until the
transaction is committed, so that if for any reason the transaction is
aborted, the mail won't be sent.

This causes problems for this worker, which does all its operations
inside a single request.

Resolve the issue by committing the current transaction after processing
each message.