Use SHA512 and perfom HKDF extract step
Nick Stenning originally had the ephemeral key set to 64 octets in
length. I lowered it to 32 when I introduced HKDF because I was
defaulting to SHA256 because it didn't need to be any longer than
the digest size according to the spec and SHA256 has a digest size
of 256 (= 32 * 8) bits.

However, Pyramid and WebOb default to using SHA512 in most cases
so it makes sense to use an input key that matches the digest size.
Some references seem to suggest using a key that's the internal
block size (which would be 1024 = 8 * 128), such as the comments
in WebOb ond Pyramid, whereas others (such as the IPSec RFC4868)
suggest or require that key lengths equal the digest size.

In any case, it seems convenient to default to SHA512 for our key
derivation and to use the digest length as the output length.

Additionally, I switched to using the HKDF class, instead of the
HKDFExpand class, in order to cover the fact that the input secret
supplied by the operator may not be cryptographically strong or may
not have well-distributed entropy.

Finally, I added a fixed, non-secret salt of 'h.security' which is
probably unnecessary with HKDF but means that if the operator is
using HKDF in other systems with the same input keying material
that the extract step will create a different master key, further
secured by the various info arguments to the derivations thereof.

I think this is probably pretty sound at this point.