Don't include streamer or notification in API app
Eventually it's possible that the streamer should become part of the
API, but the notification package should probably not, and was only
added (in b8c7863) as a temporary measure.

Remove them both for now, in the spirit of making sure that code
executes either in the main application, or in the API application, but
not both: integration between the main app and the API should eventually
be through the API or through external persistent message queues only.