Move extension build into hypothesis-buildext tool
This commit moves the machinery that builds the browser extensions into
its own module, `h.buildext`, with its own command-line tool,
`hypothesis-buildext`.

At the same time, it makes sure that the application environment is
prepared, as far as possible, in the same way for all commandline
scripts, using `pyramid.bootstrap` where possible. Hopefully, this
should make it harder to break `h.script` and `h.buildext` as a result
of changes to the rest of the application, because now the application
is loaded in the same way as gunicorn/pshell/etc will load it from the
.ini file.