Fix share dialog
Previously, the event from the initial click on the "share" event was
swallowed by the click handler on the parent thread:

    https://github.com/hypothesis/h/blob/fef9f43/h/static/scripts/directives/thread.coffee#L198-L221

Removing this has meant that the share dialog never shows.

Incidentally, this also fixes #1802, which was caused by the same code
in the thread directive, which swallowed all click events from its
children and hence prevented the closure of the share dialog.